cleanup code docs and refactor for brevity
